+++ apps/web/drizzle/0012_string-syntax.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `strings` ADD `syntax` text;